# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GnoccaTrack API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clicks/user/{deviceId}:
    get:
      operationId: getUserClicks
      tags: [clicks]
      summary: Get click history for a device/user
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User click history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Click"
...
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create an anonymous user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Nickname already taken
  /users/{deviceId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile by device ID
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/{deviceId}/verify-pin:
    post:
      operationId: verifyUserPin
      tags: [users]
      summary: Verify a user's PIN
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinVerifyInput"
      responses:
        "200":
          description: PIN verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinVerifyResult"
  /users/{deviceId}/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get user stats and badge
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
...
components:
  schemas:
...
    Click:
      type: object
      required: [id, deviceId, hotspotId, venueName, createdAt]
      properties:
        id:
          type: integer
        deviceId:
          type: string
        hotspotId:
          type: integer
        venueOsmId:
          type: ["string", "null"]
        venueName:
          type: string
        city:
          type: ["string", "null"]
        createdAt:
          type: string
...
    User:
      type: object
      required: [id, deviceId, nickname, clickCount, isBlocked, createdAt]
      properties:
        id:
          type: integer
        deviceId:
          type: string
        nickname:
          type: string
        clickCount:
          type: integer
        isBlocked:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [deviceId, nickname, pin]
      properties:
        deviceId:
          type: string
        nickname:
          type: string
        pin:
          type: string
    PinVerifyInput:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
    PinVerifyResult:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
    UserStats:
      type: object
      required: [clickCount, badge, recentClicks]
      properties:
        clickCount:
          type: integer
        badge:
          type: string
        badgeLabel:
          type: string
        recentClicks:
          type: array
          items:
            $ref: "#/components/schemas/Click"
...